requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
streamlit==1.41.1
requests==2.32.3
openpyxl==3.1.5 
python-dotenv==1.0.1
